refactor account controller
