+++ lib/views/AccountView/components/SearchModeComponent.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:Videotheque/Globals.dart';
-import 'package:Videotheque/controllers/account_controller/AccountController.dart';
+import 'package:Videotheque/controllers/AccountController/AccountController.dart';
 import 'package:community_material_icon/community_material_icon.dart';
 import 'package:flutter/material.dart';
