+++ lib/views/AccountView/components/StatsComponent.dart
@@ -1,6 +1,6 @@
 import 'package:Videotheque/Globals.dart';
 import 'package:Videotheque/components/divider_component.dart';
-import 'package:Videotheque/controllers/account_controller/AccountController.dart';
+import 'package:Videotheque/controllers/AccountController/AccountController.dart';
 import 'package:community_material_icon/community_material_icon.dart';
 import 'package:flutter/material.dart';
